requirements.txt
asgiref==3.10.0
boto3==1.40.63
botocore==1.40.63
dj-database-url==3.0.1
Django==5.2
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-storages==1.14.6
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
